third_party/WebKit/LayoutTests/fast/dom/shadow/shadowroot-clonenode-expected.txt

Calling ShadowRoot.cloneNode() should throw a DataCloneError exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.cloneNode() threw exception DataCloneError: An object could not be cloned..
PASS successfullyParsed is true

TEST COMPLETE
